refactor: split openai responses request builder
